// client/generated-api/models/OwnJournalNoteDto.ts
// tslint:disable
/**
 * Lappuzet
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import {
    JournalNoteAdditionalDto,
} from './';

/**
 * @export
 * @interface OwnJournalNoteDto
 */
export interface OwnJournalNoteDto {
    /**
     * @type {string}
     * @memberof OwnJournalNoteDto
     */
    created_at: string;
    /**
     * @type {number}
     * @memberof OwnJournalNoteDto
     */
    id: number;
    /**
     * @type {JournalNoteAdditionalDto}
     * @memberof OwnJournalNoteDto
     */
    journal_note_additional: JournalNoteAdditionalDto;
    /**
     * @type {string}
     * @memberof OwnJournalNoteDto
     */
    message: string;
    /**
     * @type {boolean}
     * @memberof OwnJournalNoteDto
     */
    resolved: boolean;
    /**
     * @type {string}
     * @memberof OwnJournalNoteDto
     */
    resolved_at: string;
}